requirements.txt
textual>=0.47.0
pyspellchecker>=0.8.0
